requirements.txt
asgiref==3.5.0
django-cors-headers==3.11.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.2
Markdown==3.3.6
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
zipp==3.7.0
git+git://github.com/erelsgl/fairpy.git